+++ crates/ego/src/solver/solver_computations.rs
@@ -93,7 +93,6 @@ where
     /// The optimized value of the criterion is returned together with the
     /// optimum location
     /// Returns (infill_obj, x_opt)
-    #[allow(dead_code)]
     #[allow(clippy::too_many_arguments)]
     //#[allow(clippy::type_complexity)]
     pub(crate) fn compute_best_point(
@@ -106,141 +105,176 @@ where
         infill_data: &InfillObjData<f64>,
         cstr_funcs: &[&(dyn ObjFn<InfillObjData<f64>> + Sync)],
         current_best: (f64, Array1<f64>),
+        actives: &Array2<usize>,
     ) -> (f64, Array1<f64>) {
         let fmin = infill_data.fmin;
 
         let mut success = false;
         let mut n_optim = 1;
         let n_max_optim = 3;
-        let mut best_point = current_best;
+        let mut best_point = current_best.to_owned();
 
         let algorithm = match self.config.infill_optimizer {
             InfillOptimizer::Slsqp => crate::optimizers::Algorithm::Slsqp,
             InfillOptimizer::Cobyla => crate::optimizers::Algorithm::Cobyla,
         };
 
-        let obj =
-            |x: &[f64], gradient: Option<&mut [f64]>, params: &mut InfillObjData<f64>| -> f64 {
-                // Defensive programming NlOpt::Cobyla may pass NaNs
-                if x.iter().any(|x| x.is_nan()) {
-                    return f64::INFINITY;
-                }
-                let InfillObjData {
-                    scale_infill_obj,
-                    scale_wb2,
-                    ..
-                } = params;
-                if let Some(grad) = gradient {
-                    // Use finite differences
-                    // let f = |x: &Vec<f64>| -> f64 {
-                    //     self.eval_infill_obj(x, obj_model, fmin, *scale_infill_obj, *scale_wb2)
-                    // };
-                    // grad[..].copy_from_slice(&x.to_vec().central_diff(&f));
+        for active in actives.outer_iter() {
+            let obj =
+                |x: &[f64], gradient: Option<&mut [f64]>, params: &mut InfillObjData<f64>| -> f64 {
+                    let InfillObjData {
+                        scale_infill_obj,
+                        scale_wb2,
+                        xbest: xcoop,
+                        ..
+                    } = params;
+                    let mut xcoop = xcoop.clone();
+                    Self::setx(&mut xcoop, &active.to_vec(), x);
 
-                    let g_infill_obj = if self.config.cstr_infill {
-                        self.eval_grad_infill_obj_with_cstrs(
-                            x,
-                            obj_model,
-                            cstr_models,
-                            cstr_tols,
-                            fmin,
-                            *scale_infill_obj,
-                            *scale_wb2,
-                        )
-                    } else {
-                        self.eval_grad_infill_obj(x, obj_model, fmin, *scale_infill_obj, *scale_wb2)
-                    };
-                    grad[..].copy_from_slice(&g_infill_obj);
-                }
-                if self.config.cstr_infill {
-                    self.eval_infill_obj(x, obj_model, fmin, *scale_infill_obj, *scale_wb2)
-                        * pofs(x, cstr_models, &cstr_tols.to_vec())
-                } else {
-                    self.eval_infill_obj(x, obj_model, fmin, *scale_infill_obj, *scale_wb2)
-                }
-            };
+                    // Defensive programming NlOpt::Cobyla may pass NaNs
+                    if xcoop.iter().any(|x| x.is_nan()) {
+                        return f64::INFINITY;
+                    }
 
-        let cstrs: Vec<_> = (0..self.config.n_cstr)
-            .map(|i| {
-                let cstr = move |x: &[f64],
-                                 gradient: Option<&mut [f64]>,
-                                 params: &mut InfillObjData<f64>|
-                      -> f64 {
-                    let scale_cstr = params.scale_cstr.as_ref().expect("constraint scaling")[i];
-                    if self.config.cstr_strategy == ConstraintStrategy::MeanConstraint {
-                        Self::mean_cstr(&*cstr_models[i], x, gradient, scale_cstr)
+                    if let Some(grad) = gradient {
+                        // Use finite differences
+                        // let f = |x: &Vec<f64>| -> f64 {
+                        //     self.eval_infill_obj(x, obj_model, fmin, *scale_infill_obj, *scale_wb2)
+                        // };
+                        // grad[..].copy_from_slice(&x.to_vec().central_diff(&f));
+
+                        let g_infill_obj = if self.config.cstr_infill {
+                            self.eval_grad_infill_obj_with_cstrs(
+                                &xcoop,
+                                obj_model,
+                                cstr_models,
+                                cstr_tols,
+                                fmin,
+                                *scale_infill_obj,
+                                *scale_wb2,
+                            )
+                        } else {
+                            self.eval_grad_infill_obj(
+                                &xcoop,
+                                obj_model,
+                                fmin,
+                                *scale_infill_obj,
+                                *scale_wb2,
+                            )
+                        };
+                        let g_infill_obj = g_infill_obj
+                            .iter()
+                            .enumerate()
+                            .filter(|(i, _)| active.to_vec().contains(i))
+                            .map(|(_, &g)| g)
+                            .collect::<Vec<_>>();
+                        grad[..].copy_from_slice(&g_infill_obj);
+                    }
+                    if self.config.cstr_infill {
+                        self.eval_infill_obj(&xcoop, obj_model, fmin, *scale_infill_obj, *scale_wb2)
+                            * pofs(x, cstr_models, &cstr_tols.to_vec())
                     } else {
-                        Self::upper_trust_bound_cstr(&*cstr_models[i], x, gradient, scale_cstr)
+                        self.eval_infill_obj(&xcoop, obj_model, fmin, *scale_infill_obj, *scale_wb2)
                     }
                 };
-                #[cfg(feature = "nlopt")]
-                {
-                    Box::new(cstr) as Box<dyn nlopt::ObjFn<InfillObjData<f64>> + Sync>
-                }
-                #[cfg(not(feature = "nlopt"))]
-                {
-                    Box::new(cstr) as Box<dyn crate::types::ObjFn<InfillObjData<f64>> + Sync>
-                }
-            })
-            .collect();
 
-        // We merge metamodelized constraints and function constraints
-        let mut cstr_refs: Vec<_> = cstrs.iter().map(|c| c.as_ref()).collect();
-        cstr_refs.extend(cstr_funcs);
+            let cstrs: Vec<_> = (0..self.config.n_cstr)
+                .map(|i| {
+                    let cstr = move |x: &[f64],
+                                     gradient: Option<&mut [f64]>,
+                                     params: &mut InfillObjData<f64>|
+                          -> f64 {
+                        let InfillObjData { xbest: xcoop, .. } = params;
+                        let mut xcoop = xcoop.clone();
+                        Self::setx(&mut xcoop, &active.to_vec(), x);
 
-        info!("Optimize infill criterion...");
-        while !success && n_optim <= n_max_optim {
-            let x_start = sampling.sample(self.config.n_start);
+                        let scale_cstr = params.scale_cstr.as_ref().expect("constraint scaling")[i];
+                        if self.config.cstr_strategy == ConstraintStrategy::MeanConstraint {
+                            Self::mean_cstr(&*cstr_models[i], &xcoop, gradient, scale_cstr)
+                        } else {
+                            Self::upper_trust_bound_cstr(
+                                &*cstr_models[i],
+                                &xcoop,
+                                gradient,
+                                scale_cstr,
+                            )
+                        }
+                    };
+                    #[cfg(feature = "nlopt")]
+                    {
+                        Box::new(cstr) as Box<dyn nlopt::ObjFn<InfillObjData<f64>> + Sync>
+                    }
+                    #[cfg(not(feature = "nlopt"))]
+                    {
+                        Box::new(cstr) as Box<dyn crate::types::ObjFn<InfillObjData<f64>> + Sync>
+                    }
+                })
+                .collect();
 
-            if let Some(seed) = lhs_optim_seed {
-                let (y_opt, x_opt) =
-                    Optimizer::new(Algorithm::Lhs, &obj, &cstr_refs, infill_data, &self.xlimits)
-                        .cstr_tol(cstr_tols.to_owned())
-                        .seed(seed)
-                        .minimize();
+            // We merge metamodelized constraints and function constraints
+            let mut cstr_refs: Vec<_> = cstrs.iter().map(|c| c.as_ref()).collect();
+            cstr_refs.extend(cstr_funcs);
 
-                info!("LHS optimization best_x {}", x_opt);
-                best_point = (y_opt, x_opt);
-                success = true;
-            } else {
-                let res = (0..self.config.n_start)
-                    .into_par_iter()
-                    .map(|i| {
-                        debug!("Begin optim {}", i);
-                        let optim_res =
-                            Optimizer::new(algorithm, &obj, &cstr_refs, infill_data, &self.xlimits)
-                                .xinit(&x_start.row(i))
-                                .max_eval(200)
-                                .ftol_rel(1e-4)
-                                .ftol_abs(1e-4)
-                                .minimize();
-                        debug!("End optim {}", i);
-                        optim_res
-                    })
-                    .reduce(
-                        || (f64::INFINITY, Array::ones((self.xlimits.nrows(),))),
-                        |a, b| if b.0 < a.0 { b } else { a },
-                    );
+            // Limits
+            let xlimits = self.xlimits.select(Axis(0), &active.to_vec());
 
-                if res.0.is_nan() || res.0.is_infinite() {
-                    success = false;
+            info!("Optimize infill criterion...");
+            while !success && n_optim <= n_max_optim {
+                let x_start = sampling.sample(self.config.n_start);
+                let x_start_coop = x_start.select(Axis(1), &active.to_vec());
+
+                if let Some(seed) = lhs_optim_seed {
+                    let (y_opt, x_opt) =
+                        Optimizer::new(Algorithm::Lhs, &obj, &cstr_refs, infill_data, &xlimits)
+                            .cstr_tol(cstr_tols.to_owned())
+                            .seed(seed)
+                            .minimize();
+
+                    info!("LHS optimization best_x {}", x_opt);
+                    best_point = (y_opt, x_opt);
+                    success = true;
                 } else {
-                    best_point = res;
+                    let res = (0..self.config.n_start)
+                        .into_par_iter()
+                        .map(|i| {
+                            debug!("Begin optim {}", i);
+                            let optim_res =
+                                Optimizer::new(algorithm, &obj, &cstr_refs, infill_data, &xlimits)
+                                    .xinit(&x_start_coop.row(i))
+                                    .max_eval(200)
+                                    .ftol_rel(1e-4)
+                                    .ftol_abs(1e-4)
+                                    .minimize();
+                            debug!("End optim {}", i);
+                            optim_res
+                        })
+                        .reduce(
+                            || (f64::INFINITY, Array::ones((xlimits.nrows(),))),
+                            |a, b| if b.0 < a.0 { b } else { a },
+                        );
+
+                    if res.0.is_nan() || res.0.is_infinite() {
+                        success = false;
+                    } else {
+                        let mut xopt_coop = best_point.1.to_vec();
+                        Self::setx(&mut xopt_coop, &active.to_vec(), &res.1.to_vec());
+                        best_point = (res.0, Array1::from(xopt_coop));
+                        success = true;
+                    }
+                }
+
+                if n_optim == n_max_optim && !success {
+                    info!("All optimizations fail => Trigger LHS optimization");
+                    let (y_opt, x_opt) =
+                        Optimizer::new(Algorithm::Lhs, &obj, &cstr_refs, infill_data, &xlimits)
+                            .minimize();
+
+                    info!("LHS optimization best_x {}", x_opt);
+                    best_point = (y_opt, x_opt);
                     success = true;
                 }
+                n_optim += 1;
             }
-
-            if n_optim == n_max_optim && !success {
-                info!("All optimizations fail => Trigger LHS optimization");
-                let (y_opt, x_opt) =
-                    Optimizer::new(Algorithm::Lhs, &obj, &cstr_refs, infill_data, &self.xlimits)
-                        .minimize();
-
-                info!("LHS optimization best_x {}", x_opt);
-                best_point = (y_opt, x_opt);
-                success = true;
-            }
-            n_optim += 1;
         }
         best_point
     }
